refactor(change-orders): extract the CO route into the JobShell panel pattern
The 1132-line ChangeOrderDetailPage — the largest page in the app — joins
the estimates/invoices shape:

- lib/changeOrderDiff.js: buildMergedRows / lineDiffTotals /
  buildDeliverableRows extracted as pure, unit-tested derivations (the
  backend's compose_change_order_diff mirrors buildMergedRows — lockstep
  note carried over).
- CODeliverablesSection.svelte owns the deliverables diff grid + inline
  drafting forms (marked as an inline-edit surface, not the future form
  kit); COLineItemsSection.svelte is a dumb renderer with actions as
  callbacks.
- ChangeOrderPanel.svelte orchestrates: CO loading, toolbar/status
  actions, modals, subnav.
- routes/jobs/JobChangeOrderPage.svelte is the thin host (job load +
  JobShell), value-keyed so a subnav coId change doesn't refetch the job.
- /jobs/:jobId/change-order/:coId now routes to the host; the old
  /change-orders/:id redirect shim is unchanged. No behavior change.

Tests: 8 new lib unit tests, host-page tests (shell chrome + no-refetch),
existing page tests migrated to ChangeOrderPanel.test.js. Full Vitest
suite green (195 files / 1228 tests); production build green.
